refactor get_user_by\2
