+++ lib/inc/accounts/UserDirectory.php
@@ -4,7 +4,7 @@
  * Github: https://github.com/JJPro
  */
 namespace nucssa_core\inc\accounts;
-use function nucssa_core\utils\{file_log};
+use function nucssa_core\utils\debug\{file_log};
 
 
 class UserDirectory {
