+++ lib/utils/debug/Logger.php
@@ -3,7 +3,7 @@
  * Author: Jason Ji
  * Github: https://github.com/JJPro
  */
-namespace nucssa_core\utils;
+namespace nucssa_core\utils\debug;
 
 class Logger
 {
